pyproject.toml
[project]
name = "pydantic-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mypy>=1.15.0",
    "ruff>=0.9.7",
    "logfire[fastapi]>=3.6.1",
    "pydantic>=2.0.0",
    "pydantic-ai[logfire]>=0.0.43",
    "pydantic-evals[logfire]>=0.0.52",
    "pydantic-settings>=2.8.1",
    "pytest>=8.0.0",
    "pytest-snapshot>=0.9.0",
    "python-dotenv>=1.0.1",
    "pytest-asyncio>=0.26.0",
    "pytest-tornasync>=0.6.0.post2",
    "pytest-trio>=0.8.0",
    "pytest-twisted>=1.14.3",
    "twisted>=24.11.0",
    "fastapi>=0.115.12",
    "sqlalchemy[asyncio,mypy]>=2.0.40",
    "asyncpg>=0.30.0",
    "uvicorn>=0.34.2",
    "aiosqlite>=0.21.0",
    "opentelemetry-instrumentation-asgi>=0.53b1",
    "typing-extensions>=4.13.2",
]

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
warn_redundant_casts = true
disallow_any_generics = true
no_implicit_reexport = true
plugins = [
    "sqlalchemy.ext.mypy.plugin",
    "pydantic.mypy"
]

[pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_any_unimported = false
warn_return_any = false

[tool.ruff]
line-length = 100

[dependency-groups]
dev = [
    "mypy-extensions>=1.1.0",
    "typing-extensions>=4.13.2",
]

[tool.pytest.ini_options]
addopts = "-s"
